fix(marketplace): escape listing JSON-LD before inlining it
The listing page inlined JSON.stringify output into a script tag, so a listing
title containing </script> could close the tag and inject markup. Listing text
is author-controlled by team managers, so this was reachable.

Jobs already had serializeJsonLd for exactly this. Moved it to lib/utils so the
marketplace can use it without importing from lib/jobs; the jobs module
re-exports it so existing callers are untouched.
